requirements.txt
asgiref==3.7.2
coverage==7.4.0
Django==4.2.8
djangorestframework==3.14.0
psycopg2==2.9.9
python-decouple==3.8
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
